src/main/java/repositories/SectionRepository.java
package repositories;

import models.*;

import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.EntityTransaction;
import javax.persistence.Persistence;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.logging.Level;
import java.util.logging.Logger;

public class SectionRepository {
    private static final EntityManagerFactory emf = Persistence.createEntityManagerFactory("library-management");
    private static final Logger logger = Logger.getGlobal();

    static {
        logger.setLevel(Level.INFO);
    }


    public static Section findById(Long id) {
        EntityManager em = null;
        Section section = null;
        try {
            em = emf.createEntityManager();
            System.out.println("Getting User from database...");
            section = em.find(Section.class, id);
            System.out.println(section.toString());
        } finally {
            if (em != null) {
                em.close();
            }
        }
        return section;
    }

    public static Section insertSection(Section section){
        EntityManager em = null;
        EntityTransaction et = null;
        try {
            em = emf.createEntityManager();
            et = em.getTransaction();
            et.begin();
            section = em.merge(section);
            et.commit();
        } catch (Exception ex) {
            if (et != null && et.isActive()) {
                logger.log(Level.SEVERE,
                        "[Canceling] Transaction with an error: {0}", ex.getMessage());
                et.rollback();
                logger.info("Canceled transaction");
            }
        } finally {
            if (em != null) {
                em.close();
            }
        }

        return section;
    }
}
